\name{binary}
\alias{binary<-}
\alias{binary.lvm}
\alias{binary<-.lvm}

\title{Binary}
\description{
binary of 
  \code{lvm} object.
}
\usage{
binary(x,...)
}
\arguments{
  \item{x}{\code{lvm}-object}
  \item{\dots}{}
}
\details{
  some details
}
\value{
  Vector of variable names.
}
\author{Klaus K. Holst}
\keyword{models}
\keyword{regression}
