% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lava.tobit-package.R
\name{lava.tobit.estimate.hook}
\alias{lava.tobit.estimate.hook}
\alias{tobit_gradient.lvm}
\alias{tobit_hessian.lvm}
\alias{tobit_logLik.lvm}
\alias{tobit_method.lvm}
\alias{tobit_objective.lvm}
\alias{tobitw_gradient.lvm}
\alias{tobitw_hessian.lvm}
\alias{tobitw_method.lvm}
\alias{lava.tobit.color.hook}
\alias{lava.tobit.init.hook}
\alias{lava.tobit.sim.hook}
\title{For internal use}
\description{
For internal use
}
\author{
Klaus K. Holst
}
\keyword{utilities}
