\name{lava.tobit.estimate.hook}
\alias{lava.tobit.color.hook}
\alias{lava.tobit.estimate.hook}
\alias{lava.tobit.init.hook}
\alias{lava.tobit.sim.hook}
\alias{tobit_gradient.lvm}
\alias{tobit_hessian.lvm}
\alias{tobit_logLik.lvm}
\alias{tobit_method.lvm}
\alias{tobit_objective.lvm}
\alias{tobitw_gradient.lvm}
\alias{tobitw_hessian.lvm}
\alias{tobitw_method.lvm}
\title{For internal use}
\description{
  For internal use
}
\author{
  Klaus K. Holst
}
\keyword{utilities}

